requirements.txt
# Core dependencies
requests>=2.31.0
httpx>=0.25.0
tenacity>=8.2.3
beautifulsoup4>=4.12.2
lxml>=4.9.3
pydantic>=2.5.0
pandas>=2.1.3
scikit-learn>=1.3.2

# NLP and entity extraction
scispacy>=0.5.3
spacy>=3.7.2
# Download models: python -m spacy download en_core_web_sm
# python -m pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_sm-0.5.3.tar.gz

# API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Graph database
neo4j>=5.14.1

# Vector search
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# PDF processing (GROBID client)
grobid-client-python>=0.0.7

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
mypy>=1.7.1
ruff>=0.1.6
black>=23.11.0

# LLM providers
openai>=1.3.0
anthropic>=0.8.0

# Additional utilities
python-dotenv>=1.0.0
tqdm>=4.66.1
